[core] Refine local KV compaction abstractions
